requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.14,<3.15
psycopg2>=2.9,<2.10
drf-spectacular>=0.25,<0.26
Pillow>=9.4,<9.5
# uwsgi>=2.0.20,<2.1